+++ cbits/souffle/IOSystem.h
@@ -17,9 +17,11 @@
 #include "RamTypes.h"
 #include "ReadStream.h"
 #include "ReadStreamCSV.h"
+#include "ReadStreamJSON.h"
 #include "SymbolTable.h"
 #include "WriteStream.h"
 #include "WriteStreamCSV.h"
+#include "WriteStreamJSON.h"
 
 #ifdef USE_SQLITE
 #include "ReadStreamSQLite.h"
@@ -77,9 +79,13 @@ private:
     IOSystem() {
         registerReadStreamFactory(std::make_shared<ReadFileCSVFactory>());
         registerReadStreamFactory(std::make_shared<ReadCinCSVFactory>());
+        registerReadStreamFactory(std::make_shared<ReadFileJSONFactory>());
+        registerReadStreamFactory(std::make_shared<ReadCinJSONFactory>());
         registerWriteStreamFactory(std::make_shared<WriteFileCSVFactory>());
         registerWriteStreamFactory(std::make_shared<WriteCoutCSVFactory>());
         registerWriteStreamFactory(std::make_shared<WriteCoutPrintSizeFactory>());
+        registerWriteStreamFactory(std::make_shared<WriteFileJSONFactory>());
+        registerWriteStreamFactory(std::make_shared<WriteCoutJSONFactory>());
 #ifdef USE_SQLITE
         registerReadStreamFactory(std::make_shared<ReadSQLiteFactory>());
         registerWriteStreamFactory(std::make_shared<WriteSQLiteFactory>());